updated load function

+++ assignment1/task2.c
@@ -55,7 +55,7 @@ int main ( int argc, char *argv[] ) {
 	}
 
 	terms = count_hashtable(hash_table);
-	load = terms / TABLE_SIZE;
+	load = (double)terms / (double)TABLE_SIZE;
 
 	printf("names.csv loaded!!!\n\n");
 	printf("Capacity = ");
